Refactor screen size handling and replace Bienvenue component with PageBienvenue and PageErreur for responsive design
