+++ src/router/index.js
@@ -3,13 +3,12 @@ import { createRouter, createWebHistory } from 'vue-router'
 import Home from '@/views/PanneauJeux.vue'
 import Recompenses from '@/views/Recompenses.vue'
 import Stats from '@/views/PageStats.vue'
-import Bienvenue from '@/views/Bienvenue.vue'
 import PageConnexion from '@/views/PageConnexion.vue'
 import PageInscription from '@/views/PageInscription.vue'
 import PageErreur from '@/views/PageErreur.vue'
 import NotFound from '@/views/NotFound.vue'
 import Leaderboard from '@/views/Leaderboard.vue'
-
+import PageBienvenue from '@/views/PageBienvenue.vue'
 import Dino from '@/views/Jeux/Dino.vue'
 import Frigo from '@/views/Jeux/Frigo.vue'
 import Dents from '@/views/Jeux/Dents.vue'
@@ -22,7 +21,7 @@ const router = createRouter({
 		{
 			path: '/',
 			name: 'root',
-			component: Bienvenue,
+			component: PageBienvenue,
 		},
 		{
 			path: '/home',
@@ -32,14 +31,14 @@ const router = createRouter({
 		},
 		{
 			path: '/erreur',
-			name: 'accueil',
-			component: Bienvenue,
+			name: 'erreur',
+			component: PageErreur,
 		},
 
 		{
 			path: '/bienvenue',
 			name: 'bienvenue',
-			component: Bienvenue,
+			component: PageBienvenue,
 		},
 		{
 			path: '/connexion',
